Add SearchCookie to front end
